Revert ""
This reverts commit ad74bf82f4ef8b34ed14cdacac683ba3841b01f2.

+++ Customers2/src/java/controller/FrontController.java
@@ -3,6 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+
 package controller;
 
 import java.io.IOException;
@@ -12,8 +13,6 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import model.Customer;
-import model.CustomerBB;
 import model.CustomerDB;
 
 /**
@@ -24,9 +23,8 @@ import model.CustomerDB;
 public class FrontController extends HttpServlet {
 
     /**
-     * Processes requests for both HTTP
-     * <code>GET</code> and
-     * <code>POST</code> methods.
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
      *
      * @param request servlet request
      * @param response servlet response
@@ -36,38 +34,27 @@ public class FrontController extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         String path = request.getServletPath();
-        CustomerDB customerDB = (CustomerDB) getServletContext().getAttribute("CustomerDB");
-        if (customerDB == null) {
-            customerDB = new CustomerDB();
-            getServletContext().setAttribute("CustomerDB", customerDB);
+                    CustomerDB customerDB = (CustomerDB) getServletContext().getAttribute("CustomerDB");
+            if (customerDB == null) {
+                customerDB = new CustomerDB();
+                getServletContext().setAttribute("CustomerDB", customerDB);
+            }
+        if ("/add.do".equals(path)) {
+            customerDB.add(request.getParameter("name"));
+            getServletContext().getRequestDispatcher("/Home").forward(request, response);
+        } else
+        if ("/delete.do".equals(path)) {
+            customerDB.delete(request.getParameter("id"));
+            getServletContext().getRequestDispatcher("/Home").forward(request, response);
+            
         }
-        switch (path) {
-            case "/add.do":
-                CustomerBB cbb = new CustomerBB(request.getParameter("name"),
-                        request.getParameter("age"));
-                Customer c = cbb.validate();
-                if (c != null) {
-                    customerDB.add(c);
-                    response.sendRedirect(getServletContext().getContextPath() + "/Home");
-                } else {
-                    request.setAttribute("customerBB", cbb);
-                    getServletContext().getRequestDispatcher("/Add").forward(request, response);
-                }
-                break;
-            case "/delete.do":
-                customerDB.delete(request.getParameter("id"));
-                response.sendRedirect(getServletContext().getContextPath() + "/Home");
-                break;
-           default: throw new ServerException(path);
-
-        }
+        throw new ServerException(path);
 
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
     /**
-     * Handles the HTTP
-     * <code>GET</code> method.
+     * Handles the HTTP <code>GET</code> method.
      *
      * @param request servlet request
      * @param response servlet response
@@ -81,8 +68,7 @@ public class FrontController extends HttpServlet {
     }
 
     /**
-     * Handles the HTTP
-     * <code>POST</code> method.
+     * Handles the HTTP <code>POST</code> method.
      *
      * @param request servlet request
      * @param response servlet response
@@ -104,4 +90,5 @@ public class FrontController extends HttpServlet {
     public String getServletInfo() {
         return "Short description";
     }// </editor-fold>
+
 }
